# orchestrator/server/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        description: Error 人类可读的错误说明
        example: insufficient permissions
        type: string
      success:
        description: Success 固定为 false
        example: false
        type: boolean
    type: object
...
host: localhost:9527
info:
  contact: {}
  description: |-
    Wingman 远程中控编排器 HTTP API（auth / RBAC / agent / script / workflow / messages / settings）。
    架构约束：Dashboard 与远程客户端只连接本 server，不直连 runtime。
  title: Wingman Orchestrator API
  version: "1.0"
paths:
...
  /agents:
    get:
      description: 返回所有已注册 agent 的状态、标签与负载数据（不含已断开条目按 registry 当前视图）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Agent 列表（状态/标签/负载）
      tags:
      - agents
  /agents/{agentId}:
    get:
      description: 返回单个 agent 的状态/标签/负载信息
      parameters:
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Agent 详情
      tags:
      - agents
  /agents/{agentId}/shutdown:
    post:
      description: 下发 system.shutdown 并将 agent 标记为 offline；需要 agents:manage 权限
      parameters:
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 关闭 Agent
      tags:
      - agents
  /agents/{agentId}/tags:
    put:
      consumes:
      - application/json
      description: 覆盖式更新 agent 标签；需要 agents:manage 权限
      parameters:
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: 标签列表
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设置 Agent 标签
      tags:
      - agents
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户名密码，返回 JWT；登录失败按 IP 限流
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 用户登录
      tags:
      - auth
  /v1/auth/logout:
    post:
      description: 无状态 JWT，客户端丢弃令牌即可；需携带有效 Bearer token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <JWT>"（由 POST /api/v1/auth/login 获取）'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"